fusb: Fix Rd detection
Previous patch prevented the Rp resistor from being
removed when it should be since Source Current
was always overridden to default. Make a change so
that it sets default only in setStateSource.

Also in DecodeCCTerminationSource if Rd is found
with default setting also test with the 1p5a
pullup to ensure it is indeed Rd.

Change-Id: Ie1c936cc4d6b38195380020862e13cbd5db5770d
Signed-off-by: Kenny Kessler <kenny.kessler@motorola.com>
Reviewed-on: https://gerrit.mot.com/1012058
SLTApproved: Slta Waiver
SME-Granted: SME Approvals Granted
Tested-by: Jira Key
Reviewed-by: Jason Knopsnyder <jasonk@motorola.com>
Reviewed-by: Ravikumar Vembu <raviv@motorola.com>
Submit-Approved: Jira Key
Tested-by: Sanjay Dhar <sanjaydhar@motorola.com>